// src/domain/model/types.ts
export type Site = 'pokerstars' | 'ggpoker' | '888' | 'ipoker' | 'wpn' | 'chico' | 'coinpoker';